Make whitespace consistent in CSS
The style.css file had a mix of tabs and spaces for indentation. Since
spaces were used in most of the file, I chose to convert the tabs to
spaces. The other way around would also be fine, as long as it's
consistent. While I was at it, I allowed Atom to strip trailing
whitespace from lines, and added a final newline.